+++ lib/stat_power.rb
@@ -3,6 +3,7 @@
 require_relative "stat_power/version"
 require_relative "stat_power/result"
 require_relative "stat_power/power_result"
+require_relative "stat_power/unequal_power_result"
 require_relative "stat_power/errors"
 require_relative "stat_power/distributions/normal"
 require_relative "stat_power/solvers/bisection"